% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_belgian_lambert_1972.R
\name{add_belgian_lambert_1972}
\alias{add_belgian_lambert_1972}
\title{Add Belgian Lambert 1972 coordinates to a data frame}
\usage{
add_belgian_lambert_1972(dat, crs = NULL)
}
\arguments{
\item{dat}{data frame with x and y coordinates}

\item{crs}{coordinate reference system of the data (defaults to \code{\link{getOption}('default_crs')})}
}
\value{
data frame with additional x_belg and y_belg columns
}
\description{
Add Belgian Lambert 1972 coordinates to a data frame
}
