% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_long_lat.R
\name{add_long_lat}
\alias{add_long_lat}
\title{Add long lat coordinates to a data frame}
\usage{
add_long_lat(dat, crs = NULL)
}
\arguments{
\item{dat}{data frame with x and y coordinates}

\item{crs}{coordinate reference system of the data (defaults to \code{\link{getOption}('default_crs')})}
}
\value{
data frame with additional long and lat columns
}
\description{
Add long lat coordinates to a data frame
}

