% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mercator.R
\name{add_mercator}
\alias{add_mercator}
\title{Add mercator coordinates to a data frame}
\usage{
add_mercator(dat, crs = NULL)
}
\arguments{
\item{dat}{data frame with x and y coordinates}

\item{crs}{coordinate reference system of the data (defaults to \code{\link{getOption}('default_crs')})}
}
\value{
data frame with additional x_merc and y_merc columns
}
\description{
Add mercator coordinates to a data frame
}

