% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addlonglat.R
\name{addlonglat}
\alias{addlonglat}
\title{Add long lat coordinates to a data frame}
\usage{
addlonglat(dat, CRS_from = CRS("+init=epsg:31370"),
  CRS_to = CRS("+proj=longlat +datum=WGS84"))
}
\arguments{
\item{dat}{data frame with x and y coordinates}

\item{CRS_from}{original CRS; defaults to '+init=epsg:31370'}

\item{CRS_to}{target CRS; defaults to '+proj=longlat +datum=WGS84'}
}
\value{
data frame with additional long and lat columns
}
\description{
Add long lat coordinates to a data frame
}

