% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back_transform.R
\name{back_transform}
\alias{back_transform}
\title{Back transformation based on data attributes}
\usage{
back_transform(dat, newdat = NULL)
}
\arguments{
\item{dat}{vector of values in transformed space, with transform attributes}

\item{newdat}{vector of values in transformed space, to be backtransformed with dat attributes}
}
\description{
For use with \code{\link{normal_transform}}
}
