% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_data_fusion.R
\name{bayesian_data_fusion}
\alias{bayesian_data_fusion}
\title{Bayesian data fusion}
\usage{
bayesian_data_fusion(data, pred, var)
}
\arguments{
\item{data}{vector of observed values}

\item{pred}{list of predictions (vectors, matrices or arrays)}

\item{var}{list of variances (vectors, matrices or arrays)}
}
\value{
list with with Bayesian data fusion results: prediction and variance (vectors, matrices or arrays)
}
\description{
Bayesian data fusion
}
