% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/belgian_lambert_1972.R
\docType{data}
\name{belgian_lambert_1972}
\alias{belgian_lambert_1972}
\title{Belgian Lambert 1972 coordinate reference system}
\format{\preformatted{Formal class 'CRS' [package "sp"] with 1 slot
  ..@ projargs: chr "+init=epsg:31370 +proj=lcc +lat_1=51.16666723333333 +lat_2=49.8333339 +lat_0=90 +lon_0=4.367486666666666 +x_0=150000.013 +y_0=5"| __truncated__
}}
\usage{
belgian_lambert_1972
}
\description{
Belgian Lambert 1972 coordinate reference system
}
\keyword{datasets}

