% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/belgian_lambert_2008.R
\name{belgian_lambert_2008}
\alias{belgian_lambert_2008}
\title{Belgian Lambert 2008 coordinate reference system}
\usage{
belgian_lambert_2008()
}
\description{
Belgian Lambert 2008 coordinate reference system
}

