% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Remove all objects from an environment}
\usage{
clear(name = .GlobalEnv)
}
\arguments{
\item{name}{name of the environment to clear; defaults to global environment}
}
\description{
Remove all objects from an environment
}

