% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{contains}
\alias{contains}
\title{Check if file contains the provided pattern}
\usage{
contains(file, pattern)
}
\arguments{
\item{file}{filename or vector of filenames}

\item{find}{text string to replace}

\item{replace}{text string to use as replacement}
}
\description{
Check if file contains the provided pattern
}

