% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_shapefile_to_dataframe.R
\name{convert_shapefile_to_dataframe}
\alias{convert_shapefile_to_dataframe}
\title{Convert a shapefile with point data to a table with the attributes plus the coordinates}
\usage{
convert_shapefile_to_dataframe(file, crs = options()$default_crs)
}
\arguments{
\item{file}{filename of the shapefile, with extension!}

\item{crs}{target crs for the x and y coordinates in the dataframe}
}
\value{
data frame with all shapefile attributes, and additional x and y columns
}
\description{
Convert a shapefile with point data to a table with the attributes plus the coordinates
}

