% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_3d_grid.R
\name{create_3d_grid}
\alias{create_3d_grid}
\title{Create a regular rectangular 3d grid coordinates data frame}
\usage{
create_3d_grid(origin, dim, width)
}
\arguments{
\item{origin}{xyz coordinates of the origin}

\item{dim}{dimensions of the grid}

\item{width}{width of the cells}
}
\value{
data frame with xyz coordinates for each grid node
}
\description{
Create a regular rectangular 3d grid coordinates data frame
}

