% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dws_gaussian_kernel.R
\name{dws_gaussian_kernel}
\alias{dws_gaussian_kernel}
\title{Distance weighted statistics: gaussian kernel}
\usage{
dws_gaussian_kernel(bandwidth, background, x)
}
\arguments{
\item{x}{distances for which to get the gaussian kernel value}
}
\value{
gaussian kernel values for distances x
}
\description{
Distance weighted statistics: gaussian kernel
}
\references{
Machuca-Mory, D. F., & Deutsch, C. V. (2012). Non-stationary Geostatistical Modeling Based on Distance Weighted Statistics and Distributions. Mathematical Geosciences, 45(1), 31–48. http://doi.org/10.1007/s11004-012-9428-z
}
