% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dws_pair_weights.R
\name{dws_pair_weights}
\alias{dws_pair_weights}
\title{Distance weighted statistics: pair weights}
\usage{
dws_pair_weights(exp_var_cloud, point_weights, t = 1)
}
\arguments{
\item{exp_var_cloud}{experimental variogram cloud}

\item{point_weights}{point weights}
}
\value{
pair weights
}
\description{
Distance weighted statistics: pair weights
}
\references{
Machuca-Mory, D. F., & Deutsch, C. V. (2012). Non-stationary Geostatistical Modeling Based on Distance Weighted Statistics and Distributions. Mathematical Geosciences, 45(1), 31–48. http://doi.org/10.1007/s11004-012-9428-z
}
