% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dws_point_weights.R
\name{dws_point_weights}
\alias{dws_point_weights}
\title{Distance weighted statistics: point weights}
\usage{
dws_point_weights(x, y, z, newX, newY, newZ, verticalExaggeration = 1,
  bandwidth, background)
}
\value{
point weights
}
\description{
Distance weighted statistics: point weights
}
\references{
Machuca-Mory, D. F., & Deutsch, C. V. (2012). Non-stationary Geostatistical Modeling Based on Distance Weighted Statistics and Distributions. Mathematical Geosciences, 45(1), 31–48. http://doi.org/10.1007/s11004-012-9428-z
}

