% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_and_replace.R
\name{find_and_replace}
\alias{find_and_replace}
\title{Find and replace text string in files}
\usage{
find_and_replace(file, find, replace, fixed = TRUE, ...)
}
\arguments{
\item{file}{filename or vector of filenames}

\item{find}{text string to replace}

\item{replace}{text string to use as replacement}
}
\description{
Find and replace text string in files
}

