% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstat_combine_vgm.R
\name{gstat_combine_vgm}
\alias{gstat_combine_vgm}
\title{Combine separately modelled isotropic horizontal and vertical variograms into a single anisotropic variogram}
\usage{
gstat_combine_vgm(horvgm, vervgm)
}
\arguments{
\item{horvgm}{the horizontal variogram}

\item{vervgm}{the vertical variogram}
}
\value{
an anisotropic gstat variogram model
}
\description{
Combine separately modelled isotropic horizontal and vertical variograms into a single anisotropic variogram
}

