% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstat_krige_2d_field.R
\name{gstat_krige_2d_field}
\alias{gstat_krige_2d_field}
\title{Krige a 2d field with gstat}
\usage{
gstat_krige_2d_field(x, y, attribute)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{attribute}{data values}
}
\value{
a matrix with the predicted values
}
\description{
Krige a 2d field with gstat
}
