% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstat_model_2d_variogram.R
\name{gstat_model_2d_variogram}
\alias{gstat_model_2d_variogram}
\title{Model a 2d variogram with gstat}
\usage{
gstat_model_2d_variogram(dat, model = "Exp")
}
\arguments{
\item{dat}{data frame with coordinates x, y and attribute z}

\item{model}{variogram model type, e.g. "Exp", "Sph", "Gau", "Mat".}
}
\value{
an isotropic gstat variogram model
}
\description{
Model a 2d variogram with gstat
}

