% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstat_remove_negative_psills.R
\name{gstat_remove_negative_psills}
\alias{gstat_remove_negative_psills}
\title{Remove negative partial sills from a gstat variogram model}
\usage{
gstat_remove_negative_psills(varmod)
}
\arguments{
\item{varmod}{variogram model}
}
\value{
variogram model without negative partial sills
}
\description{
Remove negative partial sills from a gstat variogram model
}
