% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstat_transform_to_lmc.R
\name{gstat_transform_to_lmc}
\alias{gstat_transform_to_lmc}
\title{Transform gstat object to linear model of coregionalization}
\usage{
gstat_transform_to_lmc(g, correct_diagonal = 1, ...)
}
\arguments{
\item{g}{gstat object}
}
\value{
gstat object
}
\description{
Transform gstat object to linear model of coregionalization
}
