% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_transform.R
\name{indicator_transform}
\alias{indicator_transform}
\title{Indicator transform}
\usage{
indicator_transform(dat, classes = levels(dat))
}
\arguments{
\item{dat}{factor}

\item{classes}{factor levels}
}
\value{
data frame with an indicator column for each factor level
}
\description{
Indicator transform
}

