% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/long_lat.R
\docType{data}
\name{long_lat}
\alias{long_lat}
\title{Longitude latitude coordinate reference system}
\format{\preformatted{Formal class 'CRS' [package "sp"] with 1 slot
  ..@ projargs: chr "+init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
}}
\usage{
long_lat
}
\description{
Commonly used by organizations that provide GIS data for the entire globe or many countries. CRS used by Google Earth.
}
\keyword{datasets}

