% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mercator.R
\docType{data}
\name{mercator}
\alias{mercator}
\title{Mercator coordinate reference system}
\format{\preformatted{Formal class 'CRS' [package "sp"] with 1 slot
  ..@ projargs: chr "+init=epsg:3857 +proj=merc +a=6378137 +b=6378137 +lat_ts=0.0 +lon_0=0.0 +x_0=0.0 +y_0=0 +k=1.0 +units=m +nadgrids=@null +no_def"| __truncated__
}}
\usage{
mercator
}
\description{
Used by Google Maps, Open Street Maps, Stamen Maps
}
\keyword{datasets}

