% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/points_on_circle.R
\name{points_on_circle}
\alias{points_on_circle}
\title{Get n points evenly spread on a circle}
\usage{
points_on_circle(centre_x, centre_y, radius, n)
}
\arguments{
\item{centre_x}{x coordinate of the circle centre}

\item{centre_y}{y coordinate of the circle centre}

\item{radius}{radius of the circle}

\item{n}{number of points on the circle}
}
\value{
Data frame with x and y coordinates
}
\description{
Get n points evenly spread on a circle
}

