% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prettyChar}
\alias{prettyChar}
\title{Format a character object with specified width, analogues to the \code{base} package \code{prettyNum} function.}
\usage{
prettyChar(x, width)
}
\arguments{
\item{x}{Character object.}

\item{width}{Width of the output string.}
}
\value{
Formatted character object.
}
\description{
Format a character object with specified width, analogues to the \code{base} package \code{prettyNum} function.
}

