% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{Coefficient of Determination}
\usage{
r2(x, y)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{y}{a numeric vector, matrix or data frame with compatible dimensions to \code{x}.}

\item{...}{arguments passed to \code{\link{cor}}.}
}
\description{
Coefficient of Determination
}

