% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.gslib}
\alias{read.gslib}
\title{Read a GSLIB file}
\usage{
read.gslib(file, nodatavalue = -99999)
}
\arguments{
\item{file}{Filename}

\item{nodatavalue}{Numeric value for missing data}
}
\value{
data frame
}
\description{
Read a GSLIB file
}

