% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gslib.R
\name{read_gslib}
\alias{read_gslib}
\title{Read a GSLIB file}
\usage{
read_gslib(file, nodatavalue = -99999, type = "points")
}
\arguments{
\item{file}{Filename}

\item{nodatavalue}{Numeric value for missing data}
}
\value{
data frame
}
\description{
Read a GSLIB file
}
