% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove_empty_strings}
\alias{remove_empty_strings}
\title{Remove empty elements from a vector of strings.}
\usage{
remove_empty_strings(stringArray)
}
\arguments{
\item{stringArray}{Vector of strings.}
}
\value{
Vector of strings without the empty items.
}
\description{
Remove empty elements from a vector of strings.
}

