% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strsplit0}
\alias{strsplit0}
\title{Split the Elements of a Character Vector, and remove empty elements.}
\usage{
strsplit0(...)
}
\arguments{
\item{...}{Parameters for \code{strsplit}.}
}
\value{
Output of \code{strsplit}, without empty elements.
}
\description{
Split the Elements of a Character Vector, and remove empty elements.
}

