% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.var.r
\name{weighted.var}
\alias{weighted.var}
\title{Weighted variance}
\usage{
weighted.var(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{an object containing the values whose weighted variance is to be computed}

\item{w}{a numerical vector of weights the same length as x giving the weights to use for elements of x}

\item{na.rm}{a logical value indicating whether NA values in x should be stripped before the computation proceeds}
}
\value{
a length-one numeric vector
}
\description{
Weighted variance
}
