% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.gslib}
\alias{write.gslib}
\title{Write a GSLIB file}
\usage{
write.gslib(dat, file, nodatavalue = -99999)
}
\arguments{
\item{dat}{data frame}

\item{file}{Filename}

\item{nodatavalue}{Numeric value for missing data}
}
\value{
None
}
\description{
Write a GSLIB file
}

