% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gslib.R
\name{write_gslib}
\alias{write_gslib}
\title{Write a GSLIB file}
\usage{
write_gslib(dat, file, nodatavalue = -99999)
}
\arguments{
\item{dat}{data frame}

\item{file}{Filename}

\item{nodatavalue}{Numeric value for missing data}
}
\value{
None
}
\description{
Write a GSLIB file
}

