% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addlonglat}
\alias{addlonglat}
\title{Add long lat coordinates to a data frame}
\usage{
addlonglat(dat, CRS_from = CRS("+init=epsg:31370"),
  CRS_to = CRS("+proj=longlat +datum=WGS84"))
}
\description{
Add long lat coordinates to a data frame
}

