% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addosmmerc}
\alias{addosmmerc}
\title{Add osm mercator coordinates to a data frame}
\usage{
addosmmerc(dat, CRS_from = CRS("+init=epsg:31370"),
  CRS_to = CRS("+proj=merc +a=6378137 +b=6378137 +lat_ts=0.0 +lon_0=0.0 +x_0=0.0 +y_0=0 +k=1.0 +units=m +nadgrids=@null +no_defs"))
}
\arguments{
\item{dat}{data frame with x and y coordinates}

\item{CRS_from}{original CRS; defaults to '+init=epsg:31370'}

\item{CRS_to}{target CRS; defaults to osm mercator}
}
\value{
data frame with additional Xosmmerc and Yosmmerc columns
}
\description{
Add osm mercator coordinates to a data frame
}

