% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/belgian_lambert_2008.R
\docType{data}
\name{belgian_lambert_2008}
\alias{belgian_lambert_2008}
\title{Belgian Lambert 2008 coordinate reference system}
\format{\preformatted{Formal class 'CRS' [package "sp"] with 1 slot
  ..@ projargs: chr "+init=epsg:3812 +proj=lcc +lat_1=49.83333333333334 +lat_2=51.16666666666666 +lat_0=50.797815 +lon_0=4.359215833333333 +x_0=6493"| __truncated__
}}
\usage{
belgian_lambert_2008
}
\description{
Belgian Lambert 2008 coordinate reference system
}
\keyword{datasets}

