% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/long_lat.R
\name{long_lat}
\alias{long_lat}
\title{Longitude latitude coordinate reference system}
\usage{
long_lat()
}
\description{
Commonly used by organizations that provide GIS data for the entire globe or many countries. CRS used by Google Earth.
}

