% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normal_transform.R
\name{normal_transform}
\alias{normal_transform}
\title{Normal transform}
\usage{
normal_transform(dat)
}
\arguments{
\item{dat}{vector of values to be normal transformed}
}
\value{
vector of transformed values, with attributes required for the backtransform
}
\description{
Normal transform
}

