% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_strings.R
\name{remove_empty_strings}
\alias{remove_empty_strings}
\title{Remove empty elements from a vector of strings.}
\usage{
remove_empty_strings(stringArray)
}
\arguments{
\item{stringArray}{Vector of strings.}
}
\value{
Vector of strings without the empty items.
}
\description{
Remove empty elements from a vector of strings.
}

