% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit0.R
\name{strsplit0}
\alias{strsplit0}
\title{Split the Elements of a Character Vector, and remove empty elements.}
\usage{
strsplit0(...)
}
\arguments{
\item{...}{Parameters for \code{strsplit}.}
}
\value{
Output of \code{strsplit}, without empty elements.
}
\description{
Split the Elements of a Character Vector, and remove empty elements.
}
